# packages/contracts/openapi/bodysense.v1.openapi.yaml
openapi: 3.1.0
info:
  title: BodySense Public API
  version: 1.0.0-vnext
  license:
    name: MIT
    identifier: MIT
  description: Canonical browser-facing REST contract for BodySense vNext. Public handlers and browser clients use this generated boundary as the single runtime contract.
servers:
- url: /
paths:
...
  /api/v1/treatments/proposals:
    post:
      operationId: generateTreatmentProposal
      summary: Generate a revisioned Treatment proposal from a reviewed DiagnosisAnalysis.
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TreatmentProposalRequest'
      responses:
        '201':
          description: Treatment revision proposal.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreatmentProposalResponse'
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409': {$ref: '#/components/responses/Conflict'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /api/v1/treatments/current:
    get:
      operationId: getCurrentTreatment
      summary: Read the deterministic current Treatment review projection without mutating state.
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Current treatment, or null when none exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentTreatmentResponse'
        '401': {$ref: '#/components/responses/Unauthorized'}
        '500': {$ref: '#/components/responses/InternalError'}
  /api/v1/treatments/current/review:
    post:
      operationId: reviewCurrentTreatment
      summary: Persist the latest deterministic current Treatment review state.
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Reviewed current treatment, or null when none exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentTreatmentResponse'
        '401': {$ref: '#/components/responses/Unauthorized'}
        '500': {$ref: '#/components/responses/InternalError'}
  /api/v1/treatments/revisions:
    get:
      operationId: listTreatmentRevisions
      summary: List immutable Treatment revisions for the authenticated user.
      security:
      - bearerAuth: []
      parameters:
      - name: limit
        in: query
        required: false
        schema: {type: integer, minimum: 1, maximum: 100, default: 20}
      responses:
        '200':
          description: Treatment revision timeline.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreatmentRevisionListResponse'
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '500': {$ref: '#/components/responses/InternalError'}
  /api/v1/treatments/revisions/{revisionId}:
    get:
      operationId: getTreatmentRevision
      summary: Get one immutable Treatment revision.
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/TreatmentRevisionId'
      responses:
        '200':
          description: Treatment revision.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreatmentRevision'
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalError'}
  /api/v1/treatments/revisions/{revisionId}/replay:
    post:
      operationId: replayTreatmentRevision
      summary: Replay an immutable Treatment revision historically or counterfactually.
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/TreatmentRevisionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TreatmentReplayRequest'
      responses:
        '200':
          description: Deterministic Treatment replay report.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreatmentReplayReport'
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ReplayUnavailable'}
        '502': {$ref: '#/components/responses/BadGateway'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /api/v1/treatments/revisions/{revisionId}/regression-export:
    get:
      operationId: exportTreatmentRegressionCase
      summary: Export one frozen Treatment developer regression case.
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/TreatmentRevisionId'
      responses:
        '200':
          description: Regression dataset envelope.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/JsonObject'}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ReplayUnavailable'}
        '502': {$ref: '#/components/responses/BadGateway'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /api/v1/treatments/revisions/{revisionId}/accept:
    post:
      operationId: acceptTreatmentRevision
      summary: Atomically accept one Treatment revision and project its TrainingPlan.
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/TreatmentRevisionId'
      requestBody:
        required: false
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AcceptTreatmentRevisionRequest'}
      responses:
        '200':
          description: Accepted Treatment and executable TrainingPlan projection.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AcceptTreatmentRevisionResponse'}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409': {$ref: '#/components/responses/Conflict'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /api/v1/treatments/revisions/{revisionId}/reject:
    post:
      operationId: rejectTreatmentRevision
      summary: Reject one proposed Treatment revision.
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/TreatmentRevisionId'
      responses:
        '204': {description: Treatment revision rejected.}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409': {$ref: '#/components/responses/Conflict'}
        '500': {$ref: '#/components/responses/InternalError'}
  /api/v1/outcomes:
    post:
      operationId: recordOutcome
      summary: Record an idempotent post-intervention outcome and project its BodyState effect.
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RecordOutcomeRequest'}
      responses:
        '200':
          description: Existing idempotent Outcome.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OutcomeMutationResponse'}
        '201':
          description: Newly created Outcome.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OutcomeMutationResponse'}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409': {$ref: '#/components/responses/Conflict'}
        '500': {$ref: '#/components/responses/InternalError'}
    get:
      operationId: listOutcomes
      summary: List post-intervention Outcomes.
      security:
      - bearerAuth: []
      parameters:
      - name: limit
        in: query
        required: false
        schema: {type: integer, minimum: 1, maximum: 500, default: 100}
      responses:
        '200':
          description: Outcome timeline.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OutcomeListResponse'}
        '400': {$ref: '#/components/responses/InvalidRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '500': {$ref: '#/components/responses/InternalError'}
...
components:
  parameters:
...
    TreatmentRevisionId:
      name: revisionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    JsonObject:
      type: object
      additionalProperties: true
...
    TreatmentProposalRequest:
      type: object
      additionalProperties: false
      properties:
        diagnosis_analysis_id:
          type: string
          format: uuid
        user_constraints: {$ref: '#/components/schemas/JsonObject'}
        change_reason: {type: string, maxLength: 1000}
    TreatmentProposalResponse:
      type: object
      additionalProperties: false
      required: [proposal]
      properties:
        proposal: {$ref: '#/components/schemas/TreatmentRevision'}
    CurrentTreatmentResponse:
      type: object
      additionalProperties: false
      required: [treatment]
      properties:
        treatment:
          anyOf:
          - {$ref: '#/components/schemas/Treatment'}
          - {type: 'null'}
    TreatmentRevisionListResponse:
      type: object
      additionalProperties: false
      required: [revisions]
      properties:
        revisions:
          type: array
          items: {$ref: '#/components/schemas/TreatmentRevision'}
...
    TreatmentRevision:
      type: object
      additionalProperties: false
      required:
      - id
      - treatment_id
      - revision
      - acceptance_state
      - lifecycle_state
      - source_body_state_revision
      - source_diagnosis_analysis_id
      - goal
      - duration_weeks
      - plan
      - user_constraints
      - evidence_ids
      - governance
      - agent_configuration_id
      - agent_configuration
      - execution_provenance
      - evidence_acquisition_trace
      - generation_decision_trace
      - acceptance_decision_trace
      - rollout_provenance
      - change_reason
      - created_at
      - interventions
      properties:
        id: *id003
        treatment_id: *id003
        revision: *id005
        acceptance_state:
          type: string
          enum:
          - proposed
          - accepted
          - rejected
        lifecycle_state:
          type: string
          enum:
          - active
          - review_recommended
          - paused
          - superseded
          - completed
        source_body_state_revision: *id005
        source_diagnosis_analysis_id: *id003
        goal: *id004
        duration_weeks: *id005
        plan:
          $ref: '#/components/schemas/TreatmentPlanContent'
        user_constraints:
          $ref: '#/components/schemas/JsonObject'
        evidence_ids:
          $ref: '#/components/schemas/StringArray'
        governance:
          $ref: '#/components/schemas/JsonObject'
        agent_configuration_id: *id004
        agent_configuration:
          $ref: '#/components/schemas/JsonObject'
        execution_provenance:
          $ref: '#/components/schemas/JsonObject'
        evidence_acquisition_trace:
          $ref: '#/components/schemas/JsonObject'
        generation_decision_trace:
          $ref: '#/components/schemas/JsonObject'
        acceptance_decision_trace:
          $ref: '#/components/schemas/JsonObject'
        rollout_provenance:
          $ref: '#/components/schemas/JsonObject'
        change_reason: *id004
        created_at: *id006
        accepted_at:
          type: string
          format: date-time
        interventions:
          type: array
          items:
            $ref: '#/components/schemas/Intervention'
...
  responses:
    InvalidRequest:
      description: Request body or BodyState fact is invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Authentication is missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Referenced BodyState item does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    Conflict:
      description: Requested operation conflicts with the durable resource state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    ServiceUnavailable:
      description: Required application capability is temporarily unavailable.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    InternalError:
      description: Unexpected server failure.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'